pyproject.toml
[tool.poetry]
name = "fsd-medic"
version = "0.1.0"
description = ""
authors = ["ternyavsky <ilya.ternyavsky@yandex.ru>"]
readme = "README.md"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fsd_medic.settings"

[tool.poetry.dependencies]
python = "^3.11"
amqp = "5.2.0"
asgiref = "3.7.2"
attrs = "23.2.0"
autobahn = "23.6.2"
automat = "22.10.0"
bidict = "0.22.1"
billiard = "4.2.0"
boto3 = "1.34.16"
botocore = "1.34.16"
celery = "5.3.6"
certifi = "2023.11.17"
cffi = "1.16.0"
channels = "4.0.0"
channels-redis = "4.1.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
constantly = "23.10.4"
cryptography = "41.0.7"
daphne = "4.0.0"
django = "5.0.1"
django-cors-headers = "4.3.1"
django-debug-toolbar = "4.2.0"
django-filter = "23.5"
django-loki = "0.1.4"
django-prometheus = "2.3.1"
django-redis = "5.4.0"
django-storages = "1.14.2"
django-stubs = "4.2.7"
django-stubs-ext = "4.2.7"
djangochannelsrestframework = "1.2.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
dnspython = "2.5.0"
drf-extra-fields = "3.7.0"
drf-yasg = "1.21.7"
eventlet = "0.35.0"
filetype = "1.2.0"
gevent = "23.9.1"
greenlet = "3.0.3"
gunicorn = "21.2.0"
h11 = "0.14.0"
hyperlink = "21.0.0"
idna = "3.6"
incremental = "22.10.0"
inflection = "0.5.1"
jmespath = "1.0.1"
kombu = "5.3.4"
msgpack = "1.0.7"
mysqlclient = "2.2.1"
packaging = "23.2"
pillow = "10.2.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.43"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyjwt = "2.8.0"
pymemcache = "4.0.0"
pyopenssl = "23.3.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-engineio = "4.8.2"
python-socketio = "5.11.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
redis = "5.0.1"
requests = "2.31.0"
s3transfer = "0.10.0"
service-identity = "23.1.0"
simple-websocket = "1.0.0"
six = "1.16.0"
sqlparse = "0.4.4"
twisted = "23.10.0"
txaio = "23.1.1"
types-pytz = "2024.1.0.20240203"
types-pyyaml = "6.0.12.12"
typing-extensions = "4.9.0"
tzdata = "2023.4"
uritemplate = "4.1.1"
urllib3 = "2.0.7"
vine = "5.1.0"
wcwidth = "0.2.13"
wsproto = "1.2.0"
zope-event = "5.0"
zope-interface = "6.1"
uvicorn = "^0.29.0"
pytest-django = "^4.8.0"
pytest = "^8.1.1"
django-mysql = "^4.13.0"
gcsa = "^2.3.0"
google-api-python-client = "^2.134.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
